backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
pydantic>=1.10.0
python-multipart>=0.0.6
pypdf>=3.9.0
python-docx>=0.8.11
python-dotenv>=1.0.0
deepgram>=3.3.7
